Add some categories to the desktop file
Previously, neither the desktop file nor the metainfo file listed any
categories for this game. This information is used by software centres
to list the game in the right section, and by GNOME Shell to set a good
default name if you drag two games together in the app grid to make a
folder.

Add the two best-fitting categories to the .desktop file. At the point
when the repository's AppStream data is built, these will also get
merged into the metainfo.